MJPEG codec flags moved into mfxconfig.h.in
