+++ _studio/shared/include/mfx_config.h
@@ -56,10 +56,6 @@
         //// vc1
         #define MFX_ENABLE_VC1_VIDEO_DECODE
 
-        // mjpeg
-        #define MFX_ENABLE_MJPEG_VIDEO_DECODE
-        #define MFX_ENABLE_MJPEG_VIDEO_ENCODE
-
         // vpp
         #define MFX_ENABLE_DENOISE_VIDEO_VPP
         #define MFX_ENABLE_VPP
@@ -81,8 +77,6 @@
     #endif // #if !defined(ANDROID)
 
     #if defined(AS_H264LA_PLUGIN)
-        #undef MFX_ENABLE_MJPEG_VIDEO_DECODE
-        #undef MFX_ENABLE_MJPEG_VIDEO_ENCODE
         #undef MFX_ENABLE_H264_VIDEO_FEI_ENCODE
         #undef MFX_ENABLE_VPP
     #endif
@@ -93,8 +87,6 @@
         #undef MFX_ENABLE_H264_VIDEO_ENCODE
         #undef MFX_ENABLE_MPEG2_VIDEO_ENCODE
         #undef MFX_ENABLE_VC1_VIDEO_DECODE
-        #undef MFX_ENABLE_MJPEG_VIDEO_DECODE
-        #undef MFX_ENABLE_MJPEG_VIDEO_ENCODE
         #undef MFX_ENABLE_DENOISE_VIDEO_VPP
         #undef MFX_ENABLE_VPP
         #undef MFX_ENABLE_SCENE_CHANGE_DETECTION_VPP
