+++ builder/profiles/fastboot.cmake
@@ -39,3 +39,5 @@ set( MFX_ENABLE_VP9_VIDEO_DECODE ON)
 set( MFX_ENABLE_H265_VIDEO_DECODE ON)
 set ( MFX_ENABLE_MPEG2_VIDEO_DECODE ON)
 set ( MFX_ENABLE_MPEG2_VIDEO_ENCODE OFF)
+set ( MFX_ENABLE_MJPEG_VIDEO_DECODE OFF)
+set ( MFX_ENABLE_MJPEG_VIDEO_ENCODE OFF)
