+++ mfxconfig.h.in
@@ -34,3 +34,5 @@
 #cmakedefine MFX_ENABLE_H265_VIDEO_DECODE
 #cmakedefine MFX_ENABLE_MPEG2_VIDEO_DECODE
 #cmakedefine MFX_ENABLE_MPEG2_VIDEO_ENCODE
+#cmakedefine MFX_ENABLE_MJPEG_VIDEO_DECODE
+#cmakedefine MFX_ENABLE_MJPEG_VIDEO_ENCODE
